Add: manifest.json
